refactor: separate the fetch request component
